DB minigame and Label fixes
A DB minigame was added with scoring implemented to it on a global variable so that anyone can update this global variable with Global.score. Label's became scenes so that they can be dragged into any minigame and updated with appropriate text without having to redo the template of the font.